+++ samples/ds18b20/src/main.c
@@ -46,7 +46,7 @@ int main(void)
 				LOG_ERR("Call `ctr_ds18b20_read` failed: %d", ret);
 			} else {
 				LOG_INF("Serial number: %llu / Temperature: %.2f C", serial_number,
-					temperature);
+					(double)temperature);
 			}
 		}
